Change type of identifier to double in toJson
The type is converted because the id can be higher than int support.

+++ src/Request.h
@@ -82,7 +82,7 @@ class Request {
          * Make a String json with the identifier
          * @return {"id": identifier}
         */
-        String toJson (int identifier) {
+        String toJson (double identifier) {
 
             JsonDocument json;
